### 2020-06-24 #2383 @whedon
[PRE REVIEW]: hal9001: Scalable highly adaptive lasso regression in R
**Submitting author:** @nhejazi (<a href="http://orcid.org/0000-0002-7127-2789">Nima Hejazi</a>)
**Repository:** <a href="https://github.com/tlverse/hal9001" target ="_blank">https://github.com/tlverse/hal9001</a>
**Version:** v0.2.6

nhejazi if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
