### 2020-06-25 0a8ebaa
fix appveyor build link; update pkgdown
