### 2020-07-01 b91f14b
rm reduce_basis on pred_x_basis
