### 2020-07-01 2e7fa03
pair basis_list and copy_map filtering with x_basis filtering
