+++ R/hal.R
@@ -165,6 +165,7 @@ fit_hal <- function(X,
   # make design matrix for HAL
   if (is.null(basis_list)) {
     basis_list <- enumerate_basis(X, max_degree)
+    names(basis_list) <- 1:(length(basis_list))
   }
 
   # generate a vector of col lists corresponding to the bases generated
@@ -178,12 +179,29 @@ fit_hal <- function(X,
   time_enumerate_basis <- proc.time()
 
   x_basis <- make_design_matrix(X, basis_list)
+  colnames(x_basis) <- 1:(ncol(x_basis))
   time_design_matrix <- proc.time()
 
   # catalog and eliminate duplicates
   copy_map <- make_copy_map(x_basis)
   unique_columns <- as.numeric(names(copy_map))
   x_basis <- x_basis[, unique_columns]
+  basis_list <- basis_list[unique_columns]
+
+  # bookkeeping: get end time of duplicate removal procedure
+  time_rm_duplicates <- proc.time()
+
+  # NOTE: keep only basis functions with some (or higher) proportion of 1's
+  if (!is.null(reduce_basis) && is.numeric(reduce_basis)) {
+    reduced_basis_map <- make_reduced_basis_map(x_basis, reduce_basis)
+    x_basis <- x_basis[, reduced_basis_map]
+    basis_list <- basis_list[reduced_basis_map]
+    reduced_idx <- which(colnames(x_basis) %in% names(copy_map))
+    copy_map <- copy_map[which(names(copy_map) %in% colnames(x_basis))]
+  }
+
+  # bookkeeping: get end time of basis reduction procedure
+  time_reduce_basis <- proc.time()
 
   # the HAL basis are subject to L1 penalty
   penalty_factor <- rep(1, ncol(x_basis))
@@ -200,18 +218,6 @@ fit_hal <- function(X,
     x_basis <- cbind(x_basis, X_unpenalized)
     penalty_factor <- c(penalty_factor, rep(0, ncol(X_unpenalized)))
   }
-
-  # bookkeeping: get end time of duplicate removal procedure
-  time_rm_duplicates <- proc.time()
-
-  # NOTE: keep only basis functions with some (or higher) proportion of 1's
-  if (!is.null(reduce_basis) && is.numeric(reduce_basis)) {
-    reduced_basis_map <- make_reduced_basis_map(x_basis, reduce_basis)
-    x_basis <- x_basis[, reduced_basis_map]
-  }
-
-  # bookkeeping: get end time of basis reduction procedure
-  time_reduce_basis <- proc.time()
 
   # NOTE: workaround for "Cox model not implemented for sparse x in glmnet"
   if (family == "cox") {
@@ -325,3 +331,4 @@ fit_hal <- function(X,
   class(fit) <- "hal9001"
   return(fit)
 }
+
