### 2020-07-01 906b667
test predict with actual new_data under reduce_basis filtering
