### 2020-07-01 d07e9ef
rm subsetting of basis list on copy map
