### 2020-07-01 1200e32
apply copy map in predict
