### 2020-07-01 1711ed7
test that coefs can be interpreted in terms of basis list cols/cutoffs
