+++ tests/testthat/test-interpretation.R
@@ -17,43 +17,51 @@ basis_list <- hal_fit$basis_list
 coefs <- as.vector(coef(hal_fit$hal_lasso))
 coefs_no_intercept <- coefs[-1]
 
-no.dups <- sum(unlist(lapply(hal_fit$copy_map, function(x) length(x)-1)))
+no.dups <- sum(unlist(lapply(hal_fit$copy_map, function(x) length(x) - 1)))
 
 test_that("Basis list can be paired to coefficients with copy map", {
-  expect_equal(length(basis_list)-no.dups, length(coefs_no_intercept))
+  expect_equal(length(basis_list) - no.dups, length(coefs_no_intercept))
 })
 
-coefs_summary_list <- lapply(seq_along(hal_fit$copy_map), function(copy_map_idx){
+coefs_summary_list <- lapply(seq_along(hal_fit$copy_map), function(copy_map_idx) {
   coef <- coefs_no_intercept[copy_map_idx]
 
   basis_list_idxs <- hal_fit$copy_map[[copy_map_idx]] # indices of duplicates
   basis_dups <- basis_list[basis_list_idxs]
 
-  do.call(rbind, lapply(seq_along(basis_dups), function(i){
-      data.table(
-        coef_idx = copy_map_idx+1, # coefficient index (incorporate intercept)
-        coef, # coefficient
-        basis_list_idx = basis_list_idxs[i], # basis list index
-        col_idx = basis_dups[[i]]$cols, # column number in x
-        col_cutoff = basis_dups[[i]]$cutoffs # coefficient index
-      )
-    }))
+  do.call(rbind, lapply(seq_along(basis_dups), function(i) {
+    data.table(
+      coef_idx = copy_map_idx + 1, # coefficient index (incorporate intercept)
+      coef, # coefficient
+      basis_list_idx = basis_list_idxs[i], # basis list index
+      col_idx = basis_dups[[i]]$cols, # column number in x
+      col_cutoff = basis_dups[[i]]$cutoffs # coefficient index
+    )
+  }))
 })
 
 coefs_basis_summary <- do.call(rbind, coefs_summary_list)
 
 test_that("Copy map successfully paired basis list with coefficients", {
-  expect_equal(length(unique(coefs_basis_summary$coef_idx)),
-               length(coefs_no_intercept))
-  expect_equal(length(unique(coefs_basis_summary$coef)),
-               length(unique(coefs_no_intercept)))
-  expect_equal(length(unique(coefs_basis_summary$basis_list_idx)),
-               length(basis_list))
+  expect_equal(
+    length(unique(coefs_basis_summary$coef_idx)),
+    length(coefs_no_intercept)
+  )
+  expect_equal(
+    length(unique(coefs_basis_summary$coef)),
+    length(unique(coefs_no_intercept))
+  )
+  expect_equal(
+    length(unique(coefs_basis_summary$basis_list_idx)),
+    length(basis_list)
+  )
 })
 
 # add intercept for completeness
-intercept <- data.table(coef_idx = 1, coef = coefs[1],
-                        basis_list_idx = NA, col_idx = NA, col_cutoff = NA)
+intercept <- data.table(
+  coef_idx = 1, coef = coefs[1],
+  basis_list_idx = NA, col_idx = NA, col_cutoff = NA
+)
 coefs_summary <- rbind(intercept, coefs_basis_summary)
 
 ################################################################################
@@ -66,26 +74,26 @@ basis_list <- hal_fit_reduced$basis_list
 coefs <- as.vector(coef(hal_fit_reduced$hal_lasso))
[…]
