### 2020-07-02 f23c15b
add intercept to coef tbl
