### 2020-07-06 #2383 @mikldk
@nhejazi Thank you for your patience. I am editing this submission.

The references in the paper are not all properly typeset (lower/upper case letters). Can you please fix that?

Do you have any suggestions for reviewers or should I go ahead and search for reviewers?
