### 2020-07-06 #2383 @nhejazi
Thank you for taking on editing of this submission, @mikldk.

We'll go ahead and fix the typos in the bibliography shortly. I don't have any suggestions for reviewers, probably best if you proceed with an initial selection.
