### 2020-07-16 #2383 @martinmodrak
@mikldk Hi, the package looks interesting and somewhat within my expertise, however I am currently leaving on vacation and I was also asked to review #2471 (which I find a bit closer to my area of interest and hence prefer for review). So realistically, I don't think I could start working on the review sooner than Aug 1st, maybe even a bit later.  If that's OK, I can review it. (I am also OK if you look for other reviewers who can react faster and fall back on me only if that fails).
