### 2020-07-21 #2383 @mdogucu
@mikldk, thank you for thinking of me. Due to the challenges of the COVID-19 pandemic, like JOSS, I am also currently operating in a "reduced service mode".  I unfortunately am not available for reviewing this package.
