### 2020-07-27 090f120
init summary method for hal fit objects
