+++ R/summary.R
@@ -1,0 +1,251 @@
+#' Summary of HAL fits
+#'
+#' @details Method for summarizing the coefficients of the Highly Adaptive
+#' Lasso estimator in terms of the basis functions corresponding to covariates
+#' and interactions of covariates, returned as a single S3 object of class
+#' \code{hal9001}.
+#'
+#' @param object An object of class \code{hal9001}, containing the results of
+#'  fitting the Highly Adaptive Lasso, as produced by \code{\link{fit_hal}}.
+#' @param lambda Optional \code{numeric} value of the lambda tuning
+#' parameter, for which corresponding coefficient values to be summarized.
+#' Defaults to CV-optimal value \code{lambda_star}, or the minimum value of
+#' \code{lambda_star}.
+#' @param only_nonzero_coefs A \code{logical} specifying whether the summary
+#' should include only non-zero coefficients.
+#' @param remove_redundant_duplicates A \code{logical} specifying whether the
+#' summary should remove redundant indicator basis function duplicates. If basis
+#' functions are duplicated, then one coefficient will correspond to all of the
+#' duplicates. If remove_redundant_duplicates is TRUE then the shorter basis
+#' function is retained. For example, the same coefficient may correspond to
+#' terms "I(age >= 50)", "I(age >= 50)*I(bmi >= 18)", and "I(education >= 16)",
+#' which means these basis functions all yield the same result. When
+#' remove_redundant_duplicates is TRUE, the second basis function is omitted
+#' due to the duplicated term "I(age >= 50)".
+#' @param round_cutoffs An \code{integer} indicating the number of decimal
+#' places to be used for rounding term cutoff values.
+#' @param ... Additional arguments passed to \code{summary} as necessary.
+#'
+#' @importFrom stats aggregate
+#' @importFrom data.table data.table rbindlist setorder
+#' @importFrom utils head
+#'
+#' @return A list summarizing a \code{hal9001} object's coefficients.
+#'
+#' @export
+summary.hal9001 <- function(object,
+                            lambda = NULL,
+                            only_nonzero_coefs = TRUE,
+                            remove_redundant_duplicates = TRUE,
+                            round_cutoffs = 4,
+                            ...) {
+  basis_list_idx <- coef_idx <- dup <- NULL
+  # retain coefficients corresponding to lambda
+  if (!is.null(lambda)) {
+    if (length(lambda) > 1) {
+      stop("Cannot summarize over multiple lambda.")
+    }
+    if (!(lambda %in% object$lambda_star)) {
+      if (is.null(object$glmnet_lasso)) {
+        stop(
+          "Coefficients for specified lamdba do not exist, or are not ",
+          "accessible since the fit of the lasso model was not returned ",
+          "(i.e., return_lasso was set to FALSE in hal_fit())"
+        )
+      } else {
+        if (!(lambda %in% object$glmnet_lasso$lambda)) {
+          stop("Coefficients for the specified lambda do not exist.")
+        } else {
+          lambda_idx <- which(object$glmnet_lasso$lambda == lambda)
+          coefs <- object$coefs[, lambda_idx]
+        }
+      }
+    } else {
+      lambda_idx <- which(object$lambda_star == lambda)
+      coefs <- object$coefs[, lambda_idx]
[…]
