### 2020-07-27 c3c3fdc
increase testing coverage
