### 2020-07-27 #2526 @whedon
[REVIEW]: hal9001: Scalable highly adaptive lasso regression in R
**Submitting author:** @nhejazi (<a href="http://orcid.org/0000-0002-7127-2789">Nima Hejazi</a>)
**Repository:** <a href="https://github.com/tlverse/hal9001" target ="_blank">https://github.com/tlverse/hal9001</a>
**Version:** v0.2.7
