### 2020-07-28 tlverse/hal9001#65 @rrrlw
License file completion
Consider filling in lines 634, 635, and 655 (title, author names, year) of the license.

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/2526).
