### 2020-07-28 tlverse/hal9001#66 @rrrlw
Paper references
The Eddelbuettel references should have "R" and "C++" capitalized.

Part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/2526).
