+++ src/make_hal_basis.cpp
@@ -40,7 +40,7 @@ BasisMap enumerate_basis(const NumericMatrix& X_sub,
 //' @param X_sub A subset of the columns of X, the original design matrix.
 //' @param cols An index of the columns that were reduced to by sub-setting.
 // [[Rcpp::export]]
-List make_basis_list(const NumericMatrix& X_sub, const NumericVector& cols){
+List make_basis_list(const NumericMatrix& X_sub, const NumericVector& cols, const IntegerVector& order_map){
 
   BasisMap bmap = enumerate_basis(X_sub, cols);
   List basis_list(bmap.size());
@@ -53,10 +53,17 @@ List make_basis_list(const NumericMatrix& X_sub, const NumericVector& cols){
     // List basis();
     // basis["cols"]=it->second;
     // basis["cutoffs"]=it->first;
+    NumericVector subCols = it->second;
+
+    IntegerVector order (subCols.length());
+    for(int i=0; i < subCols.length(); i++){
+      order[i] = order_map[subCols[i]-1];
+    }
 
     List basis = List::create(
       Rcpp::Named("cols") = it->second,
-      Rcpp::Named("cutoffs") = it->first
+      Rcpp::Named("cutoffs") = it->first,
+      Rcpp::Named("orders") = order
     );
     basis_list[index++] = basis;
   }
@@ -76,18 +83,28 @@ List make_basis_list(const NumericMatrix& X_sub, const NumericVector& cols){
 //' @param cutoffs Numeric providing thresholds.
 //'
 // [[Rcpp::export]]
-bool meets_basis(const NumericMatrix& X, const int row_num,
-                 const IntegerVector& cols, const NumericVector& cutoffs) {
+double meets_basis(const NumericMatrix& X, const int row_num,
+                   const IntegerVector& cols, const NumericVector& cutoffs,  const IntegerVector& orders) {
   int p = cols.length();
+  double value = 1;
+
 
   for (int i = 0; i<p; i++) {
     double obs = X(row_num,cols[i] - 1); // using 1-indexing for basis columns
-    if(!(obs > cutoffs[i])) {
-      return(false);
+    int order =  orders[i];
+    double cutoff = cutoffs[i];
+    if(!(obs > cutoff)) {
+      return(0);
     }
+    if(order!=0){
+      value = value * pow((obs - cutoff),order);
+    }
+
   }
-  return(true);
+
+  return(value);
 }
+
 
 
 //------------------------------------------------------------------------------
@@ -105,18 +122,23 @@ bool meets_basis(const NumericMatrix& X, const int row_num,
 void evaluate_basis(const List& basis, const NumericMatrix& X, SpMat& x_basis,
                     int basis_col) {
   int n = X.rows();
+
+
   //split basis into x[1] x[-1]
   //find sub-bases
   //intersect
-
   IntegerVector cols = as<IntegerVector>(basis["cols"]);
   NumericVector cutoffs = as<NumericVector>(basis["cutoffs"]);
+  IntegerVector orders =  as<IntegerVector>(basis["orders"]);
   for (int row_num = 0; row_num < n; row_num++) {
-
-    if (meets_basis(X, row_num, cols, cutoffs)) {
-      //we can add a positive indicator for this row, basis
-      x_basis.insert(row_num, basis_col) = 1;
+    double value = meets_basis(X, row_num, cols, cutoffs,  orders);
+    if (value!=0) {
+      //Add value
+      x_basis.insert(row_num, basis_col) = value;
     }
+
+
+
   }
 }
 
@@ -178,3 +200,4 @@ SpMat make_design_matrix(const NumericMatrix& X, const List& blist) {
   x_basis.makeCompressed();
   return(x_basis);
 }
+
