+++ R/predict.R
@@ -44,10 +44,13 @@ predict.hal9001 <- function(object,
 
   # generate design matrix
   pred_x_basis <- make_design_matrix(new_data, object$basis_list)
+
+  if(!is.null(object$copy_map)){
   group <- object$copy_map[[1]]
 
   # reduce matrix of basis functions
   pred_x_basis <- apply_copy_map(pred_x_basis, object$copy_map)
+  }
 
   # NOTE: keep only basis functions with some (or higher) proportion of 1's
   if (!is.null(object$reduce_basis) && is.numeric(object$reduce_basis)) {
