+++ man/enumerate_basis.Rd
@@ -4,7 +4,12 @@
 \alias{enumerate_basis}
 \title{Enumerate Basis Functions}
 \usage{
-enumerate_basis(x, max_degree = NULL)
+enumerate_basis(
+  x,
+  max_degree = NULL,
+  order_map = rep(0, ncol(x)),
+  include_zero_order = F
+)
 }
 \arguments{
 \item{x}{An input \code{matrix} containing observations and covariates
