+++ NAMESPACE
@@ -2,14 +2,17 @@
 
 S3method(predict,SL.hal9001)
 S3method(predict,hal9001)
+S3method(print,formula_hal9001)
 export(SL.hal9001)
 export(apply_copy_map)
 export(enumerate_basis)
 export(fit_hal)
+export(formula_hal9001)
 export(make_copy_map)
 export(make_design_matrix)
 export(squash_hal_fit)
 import(Rcpp)
+import(stringr)
 importFrom(Matrix,tcrossprod)
 importFrom(Rcpp,sourceCpp)
 importFrom(assertthat,assert_that)

[…]
