+++ R/formula_hal9001.R
@@ -1,0 +1,226 @@
+
+#' HAL formula: Formula for specifying functional form of HAL
+#' @details The function allows users to specify the functional form/model of
+#' hal9001 similar to in \code{\link[glm]{glm}}. The user can specify which interactions to include,
+#' monotonicity constraints, and smoothness constraints.
+#' The returned formula object can be fed directly into \code{fit_hal}
+#' and the fit can be run with minimal (no) user input.
+#' @param formula A character string specifying the hal9001 model.
+#' The format should be of the form "y ~ h(x) + i(w) + d(z) + h(x,w) + h(x,w,z) + .^2"
+#' where "y" is the outcome and "w,x,y,z" are variables in \code{data}.
+#' Each term represents a main-term/interaction to be included in the model.
+#'  For example, h(x), i(x), and d(x) each signify that all
+#'  one-way/main term basis functions of the variable x should be included.
+#'  h(x,w), i(x,w), d(x,w) specifies that all interaction (two-way) basis functions between x and w
+#'  should be included in the model. Similarly, h(x,w,z), i(x,w,z), d(x,w,z) specifies
+#'  that all interaction (three-way) basis functions between x,w,z should be included in the model.
+#'  Note that "y ~ h(x,y,z)" will only construct three-way basis functions for x,y,z
+#'  and not the two-way and one-way basis functions.
+#'  Additionally, a formula of the form "y ~ ." will generate all one-way main term basis functions for variables in \code{data}.
+#'  Similarly, "y ~ .^2" will generate all basis functions up to degree 2 for all variables in \code{data}.
+#'  More generally, "y ~ .^max_degree" will construct all basis functions up to degree max_degree.
+#'  One can combine all the notions above. For example,
+#'  "y ~ h(x,w,z) + ." and "y ~ h(x,w,z) + .^2" will generate all one-way (resp. up to two-way) basis functions
+#'  and additionally all the three-way interaction basis functions between variables w,x,z.
+#' The letters h, i, d specify functional restrictions of each term:
+#' h specifies no constraints on the term,
+#' i specifies that the term should be enforced to be monotonely increasing,
+#' d specifies that the term should be enforced to be monotonely decreasing.,
+#' @param data A data.frame or named matrix containing the outcome and covariates specified in the formula.
+#' @param smoothness_orders Same as \code{smoothness_orders} in function \code{fit_hal}.
+#' Note it should be of length 1 or length ncol(data)-1. Vector recycling will be employed otherwise.
+#' @param include_zero_order Same as \code{include_zero_order} in function \code{fit_hal}
+#' @param bins Same as \code{bins} in function \code{fit_hal}
+#' @param exclusive_dot Boolean indicator for whether the "." and ".^max_degree" arguments in the formula
+#' should be treated as exclusive or inclusive the variables already specified in the formula.
+#' For example, if "y ~ h(x,w) + ." should the "." be interpreted as: add all one-way basis functions
+#' for the variables remaining in \code{data} not yet specified in the formula (i.e. excluding x,w),
+#' or: add all one-way basis functions for all variables in the model (including x,w).
+#' As an example. if \code{exclusive_dot} is true then "y ~h(x) + .^2" and "y ~ .^2" specify the same formula, i.e. generate all basis functions up to degree 2.
[…]
