+++ R/screening_basis_hal.R
@@ -1,0 +1,324 @@
+
+
+quantizer = function(X, bins) {
+  if (is.null(bins)) {
+    return(X)
+  }
+  X = as.matrix(X)
+
+  convertColumn = function(x) {
+    quants = seq(0, 1, 1 / bins)
+    q = quantile(x, quants)
+
+    nearest <- findInterval(x, q)
+    x <- q[nearest]
+    return(x)
+  }
+  quantizer = function(X) {
+    as.matrix(apply(X, MARGIN = 2, FUN = convertColumn))
+  }
+  return(quantizer(X))
+}
+
+
+screen_basis = function(basis_list,
+                        X,
+                        y,
+                        index_to_keep = NULL,
+                        return_index = F,
+                        lower.limits = -Inf,
+                        upper.limits = Inf,
+                        screen_at_which_lambda = NULL,
+                        family = "gaussian") {
+  print(paste0("Current basis size is ", length(basis_list)))
+  X = as.matrix(X)
+  x_basis = make_design_matrix(X, basis_list)
+
+
+
+  if (!is.null(screen_at_which_lambda)) {
+    fit = glmnet::cv.glmnet(
+      x = x_basis,
+      y = y,
+      family = family,
+      lower.limits = lower.limits,
+      upper.limits = upper.limits
+    )
+    lambda = fit$lambda.min
+
+    #fit =  glmnet::glmnet(x = x_basis, y = y, family = family, lambda =  fit$lambda.min, lower.limits = lower.limits, upper.limits = upper.limits)
+
+  }
+  else{
+    fit  = glmnet::glmnet(
+      x = x_basis,
+      y = y,
+      family = family,
+      lower.limits = lower.limits,
+      upper.limits = upper.limits
+    )
+    lambda = min(fit$lambda)
+
+  }
+
+
+  if (is.null(screen_at_which_lambda)) {
+    betas = fit$beta
+    coefs = apply(as.matrix(betas), 2, function(v) {
+      as.vector(which(v != 0))
+    })
+    keep = coefs[[length(coefs)]]
+
+  }
+  else if (screen_at_which_lambda == "lambda.min") {
+    betas = stats::coef(fit, s = fit$lambda.min)
+    keep = which(betas != 0)
+
+  }
+  else if (screen_at_which_lambda == "lambda.1se") {
+    betas = stats::coef(fit, s = fit$lambda.1se)
+    keep = which(betas != 0)
+  }
+
+
+  if (length(keep) == 1) {
+    keep = union(keep, c(2, 1))
+  }
+
+  len  = length(basis_list)
+  keep = setdiff(keep, index_to_keep)
+
+
+  print(paste0("Amount of higher order basis functions added: ", length(keep)))
+  if (return_index) {
+    print(paste0(
+      "Current basis size is ",
+      length(index_to_keep) + length(keep)
+    ))
+
+    return(keep - max(index_to_keep))
+  }
+  basis_list = basis_list[keep]
+  print(paste0(
+    "Current basis size is ",
+    length(index_to_keep) + length(basis_list)
+  ))
+
+  return(basis_list)
+}
+
+merge_basis = function(reduced_basis_list1,
+                       reduced_basis_list2,
+                       X) {
+  if (length(reduced_basis_list1) == 0 | length(reduced_basis_list2)) {
+    return(NULL)
+  }
+  if (is.null(reduced_basis_list1) == 0 |
+      is.null(reduced_basis_list2)) {
+    return(NULL)
+  }
+  X = as.matrix(X)
+  len1 = length(reduced_basis_list1)
+  len2 = length(reduced_basis_list2)
+  mergeBasis = function(lst1, lst2) {
+    lst = list()
[…]
