### 2020-07-30 c5a9b25
Update formula_hal9001.R
