### 2020-07-30 66bb6bd
Added : and * functionality to formula
