+++ R/formula_hal9001.R
@@ -31,12 +31,15 @@
 #' Note it should be of length 1 or length ncol(data)-1. Vector recycling will be employed otherwise.
 #' @param include_zero_order Same as \code{include_zero_order} in function \code{fit_hal}
 #' @param bins Same as \code{bins} in function \code{fit_hal}
+#' @param generate_lower_degrees Boolean indicator for whether all lower interaction/main term basis functions should be generated for each
+#' specified term. If true then "y~h(x,w)" behaves similar to "y~x*w" in \code{formula},
+#' and if false then it behaves similar to "y~x:y" in \code{formula}.
 #' @param exclusive_dot Boolean indicator for whether the "." and ".^max_degree" arguments in the formula
 #' should be treated as exclusive or inclusive the variables already specified in the formula.
 #' For example, if "y ~ h(x,w) + ." should the "." be interpreted as: add all one-way basis functions
 #' for the variables remaining in \code{data} not yet specified in the formula (i.e. excluding x,w),
 #' or: add all one-way basis functions for all variables in the model (including x,w).
-#' As an example. if \code{exclusive_dot} is false then "y ~h(x) + .^2" and "y ~ .^2" specify the same formula, i.e. generate all basis functions up to degree 2.
+#' As an example. if \code{exclusive_dot} is false then "y ~ h(x) + .^2" and "y ~ .^2" specify the same formula, i.e. generate all basis functions up to degree 2.
 #' However, if \code{exclusive_dot} is true, then "y ~ h(x) + .^2"  encodes a different formula than "y ~ .^2".
 #' Specifically, it means to generate one way basis functions for 'x' and then all basis functions
 #' up to degree 2 for other variables excluding 'x' in \code{data}. As a result, no interactions will be added for the variable 'x'.
@@ -51,6 +54,7 @@ formula_hal9001 <-
            smoothness_orders = NULL,
            include_zero_order = F,
            bins = NULL,
+           generate_lower_degrees = F,
            exclusive_dot = F) {
     form = formula
     if (is.null(smoothness_orders) | !is.numeric(smoothness_orders)) {
@@ -64,21 +68,59 @@ formula_hal9001 <-
     order_map = smoothness_orders
     form = stringr::str_replace_all(form, " ", "")
 
-    reg = "([^\\s])~([idh]\\(([^\\s()+]+)\\)|\\.(\\^[0-9])?)(?:\\+[ihd]\\(([^\\s()]+)\\))*(\\+\\.(\\^[0-9])?)?$"
+    term_star = stringr::str_match_all(form, "([ihd]\\([^)]+\\)[*][ihd]\\([^+]+\\))")[[1]]
+    term_star = term_star[,-1]
+
+    process_star = function(term){
+      pieces = str_split(term, "[*]")[[1]]
+
+      combs = lapply(1: (length(pieces)-1), function(i){
+        combos = combn(pieces, i)
+        combos = lapply(1:ncol(combos), function(j){combos[,j]})
+
+        return(combos)
+      })
+
+      combs = unlist(combs, recursive = F)
+
+      combs = sapply(combs, function(term){paste0(term, collapse = ":")})
+      return(combs)
+    }
+    if(stringr::str_detect(form, "[*]")){
+      term_star = sapply(term_star, process_star)
+      form = paste0(form, "+", paste0(term_star, collapse = "+"))
+    }
+
+
+
+    form = stringr::str_replace_all(form, "\\)[*:][ihd]\\(", ",")
+
+
[…]
