### 2020-07-30 70b34dc
Added expanded formula
