### 2020-07-30 6c5672e
Added expanded formula
