### 2020-07-30 b073b0d
Added expanded formula
