### 2020-07-30 49be549
Added expanded formula
