+++ R/make_basis_fixed.R
@@ -1,167 +1,0 @@
-#' List Basis Functions
-#'
-#' Build a list of basis functions from a set of columns
-#'
-#' @param cols Index or indices (as \code{numeric}) of covariates (columns) of
-#'  interest in the data matrix \code{x} for which basis functions ought to be
-#'  generated. Note that basis functions for interactions of these columns are
-#'  computed automatically.
-#' @param x A \code{matrix} containing observations in the rows and covariates
-#'  in the columns. Basis functions are computed for these covariates.
-#'
-#' @return A \code{list} containing the basis functions generated from a set of
-#'  input columns.
-basis_list_cols_order <- function(cols, x, order_map, include_zero_order, include_lower_order) {
-
-
-  basis_list <- basis_list_cols(cols, x, order_map, include_zero_order, include_lower_order)
-  # Generate edge basis (including all lower order basis functions)
-  # Do not enforce that knot points are observed, as this is unlikely for edge points.
-  # Edge basis functions are important so just generate all of them from min of each column.
-  x_edge <- matrix(apply(x,2,min), nrow = 1)
-  basis_list_edge <- basis_list_cols(cols, x_edge, order_map, F, T)
-  basis_list <- c(basis_list, basis_list_edge)
-
-
-  # output
-  return(basis_list)
-}
-
-basis_list_cols <- function(cols, x, order_map, include_zero_order, include_lower_order = F) {
-  # first, subset only to columns of interest
-  x_sub <- x[, cols, drop = FALSE]
-  # call Rcpp routine to produce the list of basis functions
-
-  basis_list <- make_basis_list(x_sub, cols, order_map)
-
-
-  #Generate lower order basis functions if needed
-  #Primarily to generate lower order edge basis functions.
-  #Inefficient: duplicate basis functions
-  if(include_lower_order){
-    if(include_zero_order){
-      k_deg <- 0
-    }
-    else{
-      k_deg <- 1
-    }
-    higher_order_cols <- cols[order_map[cols] > k_deg]
-    if(length(higher_order_cols) > 0){
-      more_basis_list = lapply(higher_order_cols, function(col){
-
-        new_order_map = order_map
-        new_order_map[col] = new_order_map[col] -1
-
-        return(basis_list_cols(cols, x, new_order_map, include_zero_order,include_lower_order=T))
-      })
-
-      basis_list = union(basis_list,unlist(more_basis_list, recursive = F))
-    }
-
-
-  }
-
-  # output
-  return(basis_list)
-}
-
-################################################################################
-
-#' Compute Degree of Basis Functions
-#'
-#' Find the full list of basis functions up to a particular degree
-#'
-#' @param x An input \code{matrix} containing observations and covariates
-#'  following standard conventions in problems of statistical learning.
-#' @param degree The highest order of interaction terms for which the basis
-#'  functions ought to be generated. The default (\code{NULL}) corresponds to
-#'  generating basis functions for the full dimensionality of the input matrix.
-#'
-#' @importFrom utils combn
-#'
-#' @return A \code{list} containing  basis functions and cutoffs generated from
-#'  a set of input columns up to a particular pre-specified degree.
-basis_of_degree  <- function(x, degree, order_map, include_zero_order, include_lower_order) {
-  # get dimensionality of input matrix
-  p <- ncol(x)
-
-  # the estimation problem is not defined when the following is violated
[…]
