+++ R/predict.R
@@ -44,13 +44,10 @@ predict.hal9001 <- function(object,
 
   # generate design matrix
   pred_x_basis <- make_design_matrix(new_data, object$basis_list)
-
-  if(!is.null(object$copy_map)){
   group <- object$copy_map[[1]]
 
   # reduce matrix of basis functions
   pred_x_basis <- apply_copy_map(pred_x_basis, object$copy_map)
-  }
 
   # add unpenalized covariates
   new_unpenalized_covariates <- ifelse(
