### 2020-08-17 tlverse/hal9001#69 @daviddewhurst
Example usage
Ideally, example usage demonstrates the utility of the software in solving real-world problems. The examples using simulated data are good and helpful for the user, but it would be good to have examples that use this software with real, large datasets. If these examples do exist in the repo, then they should be more clearly labeled as they are not obvious. Maybe making an `/examples` directory would help.
