### 2020-08-17 #2526 @daviddewhurst
See completed review (except for two bullet point) and corresponding two issues on repo.
