### 2020-08-23 adb6c44
fix bibtex stylization, as per #66
