+++ paper/refs.bib
@@ -169,7 +169,7 @@ pages = {351--371}
 }
 
 @article{eddelbuettel2011rcpp,
-  title={Rcpp: Seamless R and C++ integration},
+  title={{Rcpp}: Seamless {R} and {C++} integration},
   author={Eddelbuettel, Dirk and Fran{\c{c}}ois, Romain and Allaire, J and
     Ushey, Kevin and Kou, Qiang and Russel, N and Chambers, John and Bates, D},
   journal={Journal of Statistical Software},
@@ -180,7 +180,7 @@ pages = {351--371}
 }
 
 @book{eddelbuettel2013seamless,
-  title={Seamless R and C++ integration with Rcpp},
+  title={Seamless {R} and {C++} integration with {Rcpp}},
   author={Eddelbuettel, Dirk},
   year={2013},
   publisher={Springer}
