### 2020-08-23 67bc25d
add background from #70, touch up paper, fix DOIs
