+++ paper/paper.md
@@ -24,7 +24,7 @@ affiliations:
     index: 3
   - name: Center for Computational Biology, University of California, Berkeley
     index: 4
-date: 24 June 2020
+date: 22 August 2020
 bibliography: refs.bib
 ---
 
@@ -38,8 +38,9 @@ selection tools and sensible defaults in order to improve the scalability of the
 algorithm. By building on existing `R` packages for lasso regression and
 leveraging compiled code in key internal functions, the `hal9001` `R` package
 provides a family of highly adaptive lasso estimators suitable for use in both
-modern data analysis tasks and computationally intensive statistics and machine
-learning research.
+modern large-scale data analytics and cutting-edge research efforts at the
+intersection of statistics and machine learning, including the emerging subfield
+of computational causal inference [@wong2020computational].
 
 # Background
 
@@ -48,28 +49,28 @@ of estimating complex (e.g., possibly infinite-dimensional) functional
 parameters at a near-parametric $n^{-1/3}$ rate under only relatively mild
 conditions [@vdl2017generally; @vdl2017uniform; @bibaut2019fast]. HAL requires
 that the space of the functional parameter be a subset of the set of càdlàg
-(right-hand continuous with left-hand limits) functions with sectional
-sectional variation norm bounded by a constant. In contrast to the wealth of
-data adaptive regression techniques that make strong local smoothness
-assumptions on the true form of the target functional, HAL regression's
-assumption of a finite sectional variation norm constitutes only a _global_
-smoothness assumption, making it a powerful and versatile approach. The
-`hal9001` package implements a zeroth-order HAL estimator, which constructs and
-selects (by lasso penalization) a linear combination of indicator basis
-functions to minimize the loss-specific empirical risk under the constraint that
-the $L_1$-norm of the vector of coefficients be bounded by a finite constant.
-Importantly, the estimator is formulated such that this finite constant is the
-sectional variation norm of the target functional.
+(right-hand continuous with left-hand limits) functions with sectional variation
+norm bounded by a constant. In contrast to the wealth of data adaptive
+regression techniques that make strong local smoothness assumptions on the true
+form of the target functional, HAL regression's assumption of a finite sectional
+variation norm constitutes only a _global_ smoothness assumption, making it a
+powerful and versatile approach. The `hal9001` package primarily implements a
+zeroth-order HAL estimator, which constructs and selects by lasso penalization
+a linear combination of indicator basis functions, minimizing the loss-specific
+empirical risk under the constraint that the $L_1$-norm of the resultant vector
+of coefficients be bounded by a finite constant. Importantly, the estimator is
+formulated such that this finite constant is the sectional variation norm of the
+target functional.
 
 Intuitively, construction of a HAL estimator proceeds in two steps. First,
 a design matrix composed of basis functions is generated based on the available
 set of covariates. The zeroth-order HAL makes use of indicator basis functions,
 resulting in a large, sparse matrix with binary entries; higher-order HAL
 estimators, which replace the use of indicator basis functions with splines,
-have been formulated but remain unimplemented. This representation of the target
-functional $f$ in terms of indicator basis functions partitions the support of
-$f$ into knot points, with indicator basis functions placed over subsets of the
-sections of $f$. Generally, very many basis functions are created, with an
+have been formulated, with implementation in a nascent stage. Representation of
+the target functional $f$ in terms of indicator basis functions partitions the
+support of $f$ into knot points, with such basis functions placed over subsets
+of sections of $f$. Generally, numerous basis functions are created, with an
 appropriate set of indicator bases then selected through lasso penalization.
[…]
