+++ paper/refs.bib
@@ -5,8 +5,8 @@
   publisher = {{IEEE}},
   author = {Benkeser, David and {van der Laan}, Mark J},
   title = {The Highly Adaptive Lasso Estimator},
-  booktitle = {2016 {IEEE} International Conference on Data Science and Advanced
-    Analytics ({DSAA})}
+  booktitle = {2016 {IEEE} International Conference on Data Science and
+    Advanced Analytics ({DSAA})}
 }
 
 @article{vdl2015generally,
@@ -55,7 +55,9 @@
   Title = {Doubly robust estimation in missing data and causal inference
     models},
   Volume = {61},
-  Year = {2005}
+  Year = {2005},
+  doi = {10.1111/j.1541-0420.2005.00377.x},
+  url = {https://doi.org/10.1111/j.1541-0420.2005.00377.x}
 }
 
 @article{vdl2019efficient,
@@ -72,20 +74,6 @@
   Title = {Uniform Consistency of the Highly Adaptive Lasso Estimator of
     Infinite-Dimensional Parameters},
   Year = {2017}
-}
-
-@article{ertefaie2020nonparametric,
-  doi = {},
-  url = {http://arxiv.org/abs/2005.11303},
-  year = {2020},
-  publisher = {},
-  journal = {},
-  volume = {},
-  number = {},
-  pages = {},
-  author = {Ertefaie, Ashkan and Hejazi, Nima S and {van der Laan}, Mark J},
-  title = {Nonparametric inverse probability weighted estimators based on the
-    highly adaptive lasso}
 }
 
 @manual{R,
@@ -125,25 +113,27 @@
 }
 
 @techreport{vdl2003unified,
-author = {{van der Laan}, Mark J and Dudoit, Sandrine},
-institution = {Division of Biostatistics, University of California, Berkeley},
-keywords = {Superlearner},
-number = {130},
-title = {{Unified cross-validation methodology for selection among estimators
-  and a general cross-validated adaptive epsilon-net estimator: finite sample
-  oracle inequalities and examples}},
-year = {2003}
+  author = {{van der Laan}, Mark J and Dudoit, Sandrine},
+  institution = {Division of Biostatistics, University of California,
+    Berkeley},
+  number = {130},
+  title = {{Unified cross-validation methodology for selection among estimators
+    and a general cross-validated adaptive epsilon-net estimator: finite sample
+    oracle inequalities and examples}},
+  year = {2003}
 }
 
 @article{vdv2006oracle,
-author = {{van der Vaart}, Aad W and Dudoit, Sandrine and {van der Laan}, Mark
-  J},
-title = {{Oracle inequalities for multi-fold cross validation}},
-journal = {Statistics \& Decisions},
-year = {2006},
-volume = {24},
-number = {3},
-pages = {351--371}
+  author = {{van der Vaart}, Aad W and Dudoit, Sandrine and {van der Laan},
+    Mark J},
+  title = {{Oracle inequalities for multi-fold cross validation}},
+  journal = {Statistics \& Decisions},
+  year = {2006},
+  volume = {24},
+  number = {3},
+  pages = {351--371}
+  doi = {10.1524/stnd.2006.24.3.351},
+  url = {https://doi.org/10.1524/stnd.2006.24.3.351}
 }
 
 @article{ertefaie2020nonparametric,
@@ -184,4 +174,11 @@ pages = {351--371}
   author={Eddelbuettel, Dirk},
   year={2013},
   publisher={Springer}
+}
+
+@article{wong2020computational,
+  title={Computational Causal Inference},
+  author={Wong, Jeffrey C},
+  journal={arXiv preprint arXiv:2007.10979},
+  year={2020}
 }
