### 2020-09-02 #2526 @nhejazi
Thanks for checking in @mikldk. https://github.com/tlverse/hal9001/pull/71 contains all changes related to the issues brought to my attention. The only outstanding issue afaik is https://github.com/tlverse/hal9001/issues/69, which recommends a few real data analysis examples. I don't have any prepared offhand, so I may ask that this be deferred as something we should add in the future rather than for the JOSS review (any input from @daviddewhurst also welcome on this). I think it's a good suggestion but, in the meantime, we could, e.g., point to data analysis sections of existing academic manuscripts.
