### 2020-09-03 #2526 @mikldk
@nhejazi I think it would be a great idea to refer to a few places where HAL is used - both in `paper.md` and in the help pages of the software.
