### 2020-09-11 #2526 @rrrlw
Apologies for the delay with this review - overall, this looks like a solid submission. The (minor) issues that I opened were addressed and closed. I think this will be a useful addition to the JOSS literature. I have left the "Examples" box unchecked until it's addressed (looks like pointing to data analysis sections of existing manuscripts (ideally ones that have the source code available?) in paper.md and in the help pages will work).

As such, once the "Examples" box is addressed, I feel comfortable recommending this package + manuscript be accepted to JOSS.

I do have an optional suggestion on how the package could be improved - a vignette on using it with tidymodels. Although not everyone subscribes to the tidymodels way of doing things, its use has grown over time and it might be worth adding an example or two showing how tidymodels users could take advantage of hal9001. Other than that, I think this is a solid package, as is evidenced by its use in multiple academic manuscripts. Thank you to the authors for this valuable contribution!
