### 2020-09-15 #2526 @nhejazi
Thanks @mikldk and @rrrlw, I'll add a few references in an _Applications_ section of the paper draft and add code from one such project that I was involved in. This will cover how to use HAL in constructing a popular inverse probability weighted estimator that is prominent in the causal inference literature. I'll ping here again when that's done so that we can discuss any lingering details to finish up the review.

@rrrlw, thanks for noting a possible tidymodels integration. I like this idea and have some plans to modularize hal9001 (in a subsequent major release). If you don't mind opening up an issue about this, I'll keep track of it there and try to make these changes at the same time.
