+++ paper/refs.bib
@@ -131,7 +131,7 @@
   year = {2006},
   volume = {24},
   number = {3},
-  pages = {351--371}
+  pages = {351--371},
   doi = {10.1524/stnd.2006.24.3.351},
   url = {https://doi.org/10.1524/stnd.2006.24.3.351}
 }
@@ -181,4 +181,94 @@
   author={Wong, Jeffrey C},
   journal={arXiv preprint arXiv:2007.10979},
   year={2020}
+}
+
+@article{ju2020robust,
+  title={Robust inference on the average treatment effect using the outcome
+    highly adaptive lasso},
+  author={Ju, Cheng and Benkeser, David and {van der Laan}, Mark J},
+  journal={Biometrics},
+  volume={76},
+  number={1},
+  pages={109--118},
+  year={2020},
+  publisher={Wiley Online Library}
+}
+
+@article{benkeser2020nonparametric,
+  title={A nonparametric super-efficient estimator of the average treatment
+    effect},
+  author={Benkeser, David and Cai, Weixin and {van der Laan}, Mark J},
+  journal={Statistical Science},
+  volume={35},
+  number={3},
+  pages={484--495},
+  year={2020},
+  publisher={Institute of Mathematical Statistics}
+}
+
+@article{diaz2020causal,
+  title={Causal mediation analysis for stochastic interventions},
+  author={D{\'\i}az, Iv{\'a}n and Hejazi, Nima S},
+  year={2020},
+  url = {https://doi.org/10.1111/rssb.12362},
+  doi = {10.1111/rssb.12362},
+  journal={Journal of the Royal Statistical Society: Series B (Statistical
+    Methodology)},
+  volume = {82},
+  number = {3},
+  pages = {661-683},
+  publisher={Wiley Online Library}
+}
+
+@article{hejazi2020efficient,
+  title = {Efficient nonparametric inference on the effects of stochastic
+    interventions under two-phase sampling, with applications to vaccine
+    efficacy trials},
+  author = {Hejazi, Nima S and {van der Laan}, Mark J and Janes, Holly E and
+    Gilbert, Peter B and Benkeser, David C},
+  year={2020},
+  url = {http://arxiv.org/abs/2003.13771},
+  doi = {},
+  journal = {Biometrics (Methodology)},
+  volume={},
+  number={},
+  pages={},
+  publisher = {Wiley Online Library}
+}
+
+@article{shortreed2017outcome,
+  title={Outcome-adaptive lasso: Variable selection for causal inference},
+  author={Shortreed, Susan M and Ertefaie, Ashkan},
+  journal={Biometrics},
+  volume={73},
+  number={4},
+  pages={1111--1122},
+  year={2017},
+  publisher={Wiley Online Library}
+}
+
+@article{hernan2020causal,
+  title={Causal inference: what if},
+  author={Hern{\'a}n, Miguel A and Robins, James M},
+  journal={Boca Raton: Chapman \& Hill/CRC},
+  year={2020}
+}
+
+@article{hejazi2020txshift-joss,
+  author = {Hejazi, Nima S and Benkeser, David C},
+  title = {{txshift}: Efficient estimation of the causal effects of
+    stochastic interventions in {R}},
+  year  = {2020},
+  journal = {under review at Journal of Open Source Software},
+  publisher = {The Open Journal}
+}
+
+@software{hejazi2020txshift-rpkg,
+  author = {Hejazi, Nima S and Benkeser, David C},
+  title = {{txshift}: Efficient Estimation of the Causal Effects of
+    Stochastic Interventions},
+  year  = {2020},
+  url = {https://github.com/nhejazi/txshift},
+  note = {R package version 0.3.4}
 }
