### 2020-09-17 #2526 @nhejazi
I've just added an _Applications_ section to the JOSS paper draft (https://github.com/tlverse/hal9001/pull/71/commits/640e773b4e5e326f854dbe2018bed757f504ec95), which briefly reviews and details the use of HAL regression in four very recent papers. I've tried to go through HAL was used in each paper and point out what advances were made by relying upon the `hal9001` package. While this doesn't contribute code examples, my hope is that interested readers will be able to check out the code released with each of these papers (as such code becomes available) so that they can review how they might use HAL in their own work. Does this seem sufficient for the examples requirement?
