### 2020-09-17 31ff139
update vignette + rebuild pkgdown
