### 2020-09-17 #2526 @rrrlw
The revisions look sufficient to me, just checked off the last 2 boxes. Will open a tidymodels issue shortly, but hal9001 is good to go on my end!
