### 2020-09-17 #2526 @nhejazi
Great! I've merged the PR related to these changes into `master`. Could we update the version in the first comment on this issue thread to 0.2.7 (instead of 0.2.6)? (I don't have permissions to do that.) Also, the PDF built from `master` should match the updated version of the paper now; this wasn't the case before the relevant PR was merged last night.
