### 2020-09-18 #2526 @mikldk
@rrrlw, @daviddewhurst: Thank you for your reviews!

@nhejazi :

* I changed version to v0.2.7
* Please check above DOI problems
* Please make a tagged release and archive (e.g. with [Zenodo](https://zenodo.org/)) as described [here](<https://joss.readthedocs.io/en/latest/submitting.html#the-review-process>), and report the version number and archive DOI in this thread. Please verify that the archive deposit has the correct metadata (title and author list), or edit these if that is not the case.
