+++ paper/paper.md
@@ -8,7 +8,7 @@ tags:
 authors:
   - name: Nima S. Hejazi
     orcid: 0000-0002-7127-2789
-    affiliation: 1, 4
+    affiliation: 1, 2, 4
   - name: Jeremy R. Coyle
     orcid: 0000-0002-9874-6649
     affiliation: 2
@@ -18,47 +18,47 @@ authors:
 affiliations:
   - name: Graduate Group in Biostatistics, University of California, Berkeley
     index: 1
-  - name: Division of Epidemiology & Biostatistics, School of Public Health, University of California, Berkeley
+  - name: Division of Biostatistics, School of Public Health, University of California, Berkeley
     index: 2
   - name: Department of Statistics, University of California, Berkeley
     index: 3
   - name: Center for Computational Biology, University of California, Berkeley
     index: 4
-date: 16 September 2020
+date: 18 September 2020
 bibliography: refs.bib
 ---
 
 # Summary
 
-The `hal9001` `R` package provides an efficient implementation of the _highly
-adaptive lasso_ (HAL), a flexible nonparametric regression and machine learning
-algorithm endowed with several theoretically convenient properties. `hal9001`
-pairs an implementation of this estimator with an array of practical variable
-selection tools and sensible defaults in order to improve the scalability of the
-algorithm. By building on existing `R` packages for lasso regression and
-leveraging compiled code in key internal functions, the `hal9001` `R` package
-provides a family of highly adaptive lasso estimators suitable for use in both
-modern large-scale data analytics and cutting-edge research efforts at the
-intersection of statistics and machine learning, including the emerging subfield
-of computational causal inference [@wong2020computational].
+The `hal9001` `R` package provides a computationally efficient implementation of
+the _highly adaptive lasso_ (HAL), a flexible nonparametric regression and
+machine learning algorithm endowed with several theoretically convenient
+properties. `hal9001` pairs an implementation of this estimator with an array of
+practical variable selection tools and sensible defaults in order to improve the
+scalability of the algorithm. By building on existing `R` packages for lasso
+regression and leveraging compiled code in key internal functions, the `hal9001`
+`R` package provides a family of highly adaptive lasso estimators suitable for
+use in both modern large-scale data analysis and cutting-edge research efforts
+at the intersection of statistics and machine learning, including the emerging
+subfield of computational causal inference [@wong2020computational].
 
 # Background
 
 The highly adaptive lasso (HAL) is a nonparametric regression function capable
 of estimating complex (e.g., possibly infinite-dimensional) functional
-parameters at a near-parametric $n^{-1/3}$ rate under only relatively mild
-conditions [@vdl2017generally; @vdl2017uniform; @bibaut2019fast]. HAL requires
-that the space of the functional parameter be a subset of the set of càdlàg
-(right-hand continuous with left-hand limits) functions with sectional variation
-norm bounded by a constant. In contrast to the wealth of data adaptive
-regression techniques that make strong local smoothness assumptions on the true
-form of the target functional, HAL regression's assumption of a finite sectional
-variation norm constitutes only a _global_ smoothness assumption, making it a
-powerful and versatile approach. The `hal9001` package primarily implements a
-zeroth-order HAL estimator, which constructs and selects by lasso penalization
-a linear combination of indicator basis functions, minimizing the loss-specific
-empirical risk under the constraint that the $L_1$-norm of the resultant vector
-of coefficients be bounded by a finite constant. Importantly, the estimator is
+parameters at a fast $n^{-1/3}$ rate under only relatively mild conditions
+[@vdl2017generally; @vdl2017uniform; @bibaut2019fast]. HAL requires that the
+space of the functional parameter be a subset of the set of càdlàg (right-hand
[…]
