+++ paper/refs.bib
@@ -192,6 +192,8 @@
   number={1},
   pages={109--118},
   year={2020},
+  url = {https://doi.org/10.1111/biom.13121},
+  doi = {10.1111/biom.13121},
   publisher={Wiley Online Library}
 }
 
@@ -245,6 +247,8 @@
   number={4},
   pages={1111--1122},
   year={2017},
+  url = {https://doi.org/10.1111/biom.12679},
+  doi = {10.1111/biom.12679},
   publisher={Wiley Online Library}
 }
 
