### 2020-09-18 bb70804
update news to 0.2.7
