### 2020-09-18 #2526 @nhejazi
Echoing @mikldk, thanks very much for the thorough reviews, @rrrlw and @daviddewhurst! Addressing the issues you raised significantly improved the `hal9001` package and JOSS paper.

@mikldk, I've made minor touch-ups throughout the paper, created a new minor release (https://github.com/tlverse/hal9001/releases/tag/v0.2.7), and added the resultant release to Zenondo (DOI: https://doi.org/10.5281/zenodo.4037393).
