### 2020-09-21 #2526 @nhejazi
Thanks for catching this @mikldk. The title of the Zenodo archive has been fixed.
