### 2020-09-25 #2526 @nhejazi
Hi @mikldk, I can address the missing DOI 10.1111/biom.13375 if necessary (the paper was _just accepted_ this week), but I don't think it would be a problem to accept as is, since the two submissions were ongoing concurrently. Happy to address this if it's an issue. Thanks for your work in curating/editing this JOSS submission.
