### 2020-09-25 13ebee9
minor changes to finalize joss paper
