+++ paper/paper.md
@@ -24,7 +24,7 @@ affiliations:
     index: 3
   - name: Center for Computational Biology, University of California, Berkeley
     index: 4
-date: 18 September 2020
+date: 25 September 2020
 bibliography: refs.bib
 ---
 
