+++ paper/refs.bib
@@ -230,9 +230,9 @@
   author = {Hejazi, Nima S and {van der Laan}, Mark J and Janes, Holly E and
     Gilbert, Peter B and Benkeser, David C},
   year={2020},
-  url = {http://arxiv.org/abs/2003.13771},
-  doi = {},
-  journal = {Biometrics (Methodology)},
+  url = {https://doi.org/10.1111/biom.13375},
+  doi = {10.1111/biom.13375},
+  journal = {Biometrics},
   volume={},
   number={},
   pages={},
